+++ src/components/FormCards/FormCards.test.tsx
@@ -1,22 +1,22 @@
 import { describe, expect, test } from 'vitest';
 import { render } from '@testing-library/react';
-import Cards from './Cards';
+import FormCards from './FormCards';
 import { MOCK_CARDS } from '../../assets/mocks';
 
-describe('<Cards> component', () => {
+describe('<FormCards> component', () => {
   const data = MOCK_CARDS;
 
   test('Shoud render', () => {
-    render(<Cards data={data} />);
+    render(<FormCards data={data} />);
   });
 
   test('Shoud render UL list', () => {
-    const { container } = render(<Cards data={data} />);
+    const { container } = render(<FormCards data={data} />);
     expect(container.firstChild?.nodeName.toUpperCase()).toEqual('UL');
   });
 
   test('Shoud have 2 cards', () => {
-    const { container } = render(<Cards data={data} />);
+    const { container } = render(<FormCards data={data} />);
     expect(container.firstChild?.childNodes.length).toEqual(data.length);
   });
 });
