+++ src/components/FormCards/FormCards.tsx
@@ -1,0 +1,17 @@
+import { PropertyData } from '../../types';
+import FormCard from '../FormCard/FormCard';
+import styles from './FormCards.module.scss';
+
+interface FormCardsProps {
+  data: PropertyData[];
+}
+
+export default function FormCards({ data }: FormCardsProps) {
+  return (
+    <ul className={styles.FormCards}>
+      {data.map((data) => (
+        <FormCard key={data.id} data={data} />
+      ))}
+    </ul>
+  );
+}
