+++ src/routes/AddProperty.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Cards from '../components/Cards/Cards';
+import FormCards from '../components/FormCards/FormCards';
 import PropertyForm from '../components/PropertyForm/PropertyForm';
 import { EmptyProps, PropertyData } from '../types';
 
@@ -24,7 +24,7 @@ export default class AddProperty extends React.Component<EmptyProps, AddProperty
     return (
       <div style={{ display: 'flex', flexDirection: 'column', gap: '2rem' }}>
         <PropertyForm onSubmit={this.handleSubmit} />
-        <Cards data={this.state.cardData} />
+        <FormCards data={this.state.cardData} />
       </div>
     );
   }
